[flutter_releases] Flutter 1.22.1 Engine Cherrypicks (#21669)
* Update Dart to 2.10.1

* Extract the WindowInsetsAnimation.Callback subclass into a separate class that will be lazily loaded (#21548)

WindowInsetsAnimation.Callback was introduced in API level 30.  This
PR moves the text input plugin's WindowInsetsAnimation.Callback subclass
into a class that will only be loaded if the embedding has checked for a
sufficient API level.

See https://github.com/flutter/flutter/issues/66908

* iOS: only add explicit transactions when there are platform views (only on main threads) (#21526)

* update licenses golden

* pin framework for build tests
