+++ DEPS
@@ -34,7 +34,7 @@ vars = {
   # Dart is: https://github.com/dart-lang/sdk/blob/master/DEPS.
   # You can use //tools/dart/create_updated_flutter_deps.py to produce
   # updated revision list of existing dependencies.
-  'dart_revision': '41eab9b49ccce8960f71c657861dc629f96295af',
+  'dart_revision': 'efd753621946a89008b76b76d85d54d1aa57fce8',
 
   # WARNING: DO NOT EDIT MANUALLY
   # The lines between blank lines above and below are generated by a script. See create_updated_flutter_deps.py
