+++ ci/licenses_golden/licenses_third_party
@@ -1,4 +1,4 @@
-Signature: 52ed6d65d7e96daef749ee003a3463a0
+Signature: fcda8704ef65a785787b4045cbe4f432
 
 UNUSED LICENSES:
