+++ shell/platform/android/test/io/flutter/plugin/editing/TextInputPluginTest.java
@@ -652,8 +652,7 @@ public class TextInputPluginTest {
     TextInputChannel textInputChannel = new TextInputChannel(mock(DartExecutor.class));
     TextInputPlugin textInputPlugin =
         new TextInputPlugin(testView, textInputChannel, mock(PlatformViewsController.class));
-    TextInputPlugin.ImeSyncDeferringInsetsCallback imeSyncCallback =
-        textInputPlugin.getImeSyncCallback();
+    ImeSyncDeferringInsetsCallback imeSyncCallback = textInputPlugin.getImeSyncCallback();
     FlutterEngine flutterEngine =
         spy(new FlutterEngine(RuntimeEnvironment.application, mockFlutterLoader, mockFlutterJni));
     FlutterRenderer flutterRenderer = spy(new FlutterRenderer(mockFlutterJni));
